MD5-CRACKER + PassList.txt
so it's an md5 cracker and yeah i put a free simple wordlist on it so you can use it. if u wanna add another don't worry just drag and drop it on the folder you installed and if u want you can replace it with PassList.txt. Have fun cracking passwords! 
JUST USE FOR ETHICAL HACKING.